Split the minimal install into conda and pip through .bat file and move imports to remove soft_import warnings during inversion.
