+++ geoapps/__init__.py
@@ -6,7 +6,3 @@
 #  (see LICENSE file at the root of this source code package).
 
 __version__ = "0.8.0"
-
-from .base.application import BaseApplication
-from .base.plot import PlotSelection2D
-from .base.selection import ObjectDataSelection
