+++ geoapps/octree_creation/driver.py
@@ -21,7 +21,6 @@ from discretize.utils import mesh_builder_xyz, refine_tree_xyz
 from geoh5py.objects import ObjectBase
 from geoh5py.ui_json import InputFile
 
-from geoapps.base.application import BaseApplication
 from geoapps.driver_base.utils import treemesh_2_octree
 
 from .params import OctreeParams
@@ -88,6 +87,8 @@ class OctreeDriver:
         if self.params.monitoring_directory is not None and os.path.exists(
             self.params.monitoring_directory
         ):
+            from geoapps.base.application import BaseApplication
+
             BaseApplication.live_link_output(self.params.monitoring_directory, octree)
 
         print(
