+++ setup.py
@@ -22,7 +22,7 @@ CLASSIFIERS = [
 
 setup(
     name="geoapps",
-    version="0.7.0",
+    version="0.8.0",
     packages=find_packages(),
     install_requires=["numpy", "h5py", "scipy", "geoh5py", "requests"],
     author="Mira Geoscience",
